src/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
coverage==5.5
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.5
django-debug-toolbar==3.2.2
django-filter==2.4.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
djangorestframework==3.12.4
idna==3.2
oauthlib==3.1.1
Pillow==8.3.1
psycopg2==2.9.1
pycparser==2.20
pygeoip==0.2.7
PyJWT==2.1.0
python-dotenv==0.19.0
python-http-client==3.3.2
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
sendgrid==6.8.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.1
urllib3==1.26.6
